Issue 14: Organized Tests A.2 - Move unit tests to separate files (Part 1)
Moved unit tests from integration test files to dedicated unit test files:

✅ test__path_parser.py - Path expression parsing unit tests
✅ test__transformations.py - Transformation function unit tests
✅ test__node_model.py - Node model unit tests
✅ test__renderer_basic.py - Basic rendering unit tests
✅ test__renderer_indentation.py - Indentation unit tests

Changes:
- Created 5 new unit test files with test__ prefix
- Updated integration files to note where tests were moved
- All 528 tests still pass
- No functional changes, just moved existing tests
